==> 1_code/00_hetnet_design_and_build.ipynb <==
import os
import json
import pandas as pd

## Define Nodes

node_info = {}

# There seem to be many types of sequence variants,
# But the edges that link them are pretty unique, so will define based on
# Those edges rather than a Q-ID
node_info['Sequence Variant'] = {
    'abbrev': 'V',
    'id': None}

node_info['Binding Site'] = {
    'abbrev': 'BS',
    'id': 'Q616005'}

node_info['Active Site'] = {
    'abbrev': 'AS',
    'id': 'Q423026'}

node_info['Super-Secondary Structure'] = {
    'abbrev': 'SS',
    'id': 'Q7644128'}

node_info['Chemical Hazard'] = {
    'abbrev': 'CH',
    'id': 'Q21167512'}

node_info['Molecular Function'] = {
    'abbrev': 'MF',
    'id': 'Q14860489'}
    
node_info['Pharmaceutical Product'] = {
    'abbrev': 'PP',
    'id': 'Q28885102'}

node_info['Structural Motif'] = {
    'abbrev': 'SM',
    'id': 'Q3273544'}

node_info['Protein Family'] = {
    'abbrev': 'PF',
    'id': 'Q417841'}

node_info['Compound'] = {
    'abbrev': 'C',
    'id': 'Q11173'}

node_info['Disease'] = {
    'abbrev': 'D',
    'id': 'Q12136'}

## All proteins are encoded by genes, so qualify with 'gene encodes protein'
node_info['Protein'] = {
    'abbrev': 'P',
    'id': 'Q8054',
    'qualifiers': [{'s': 'Gene', 'p': 'P688', 'o': 'Protein', 'not': False},
                   {'s': 'Gene', 'p': 'P703', 'o': 'Q15978631', 'not': False}]}

## Genes will be defined as a human gene that does not encode a ANYTHING
## None in Object means it only cares about P688.... (if you specifiy proteins, get memory erros...)
node_info['Gene'] = {
    'abbrev': 'G',
    'id': 'Q7187',
    'qualifiers': [{'s': 'Gene', 'p': 'P688', 'o': None, 'not': True}, 
                   {'s': 'Gene', 'p': 'P703', 'o': 'Q15978631', 'not': False}]}
    
node_info['Protein Domain'] = {
    'abbrev': 'PD',
    'id': 'Q898273'}
    
node_info['Cellular Component'] = {
    'abbrev': 'CC',
    'id': 'Q5058355'}
    
node_info['Biological Pathway'] = {
    'abbrev': 'PW',
    'id': 'Q4915012'}

node_info['Biological Process'] = {
    'abbrev': 'BP',
    'id': 'Q2996394'}

# Don't want symptoms that could also be diseases....
# Get a lot of overlapping types espectially on chemcial hazard edges
node_info['Symptom'] = {
    'abbrev': 'S',
    'id': 'Q169872',
    'qualifiers': [{'s': 'Symptom', 'p': 'P31', 'o': 'Q12136', 'not': True},
                   {'s': 'Symptom', 'p': 'P279', 'o': 'Q12136', 'not': True}]}
    
node_info['Medical Specialty'] = {
    'abbrev': 'MS',
    'id': 'Q930752'}

# Chemical Role doesn't exist, but is an important type....
# Again defining based on the edge
node_info['Chemical Role'] = {
    'abbrev': 'CR',
    'id': None}
    
node_info['Anatomical Structure'] = {
    'abbrev': 'A',
    'id': 'Q4936952'}

edge_info = {}

edge_info['BIOLOGICAL_PROCESS'] = {
    'abbrev': 'bp',
    'id': 'P682'}

edge_info['VARIANT_OF'] = {
    'abbrev': 'v',
    'id': 'P3433'}

edge_info['CELL_COMPONENT'] = {
    'abbrev': 'cc',
    'id': 'P681'}

edge_info['TREATS'] = {
    'abbrev': 't',
    'id': 'P2175',
    'rev_id': 'P2176'}

edge_info['GENETIC_ASSOCIATION'] = {
    'abbrev': 'a',
    'id': 'P2293'}

edge_info['HAS_ACTIVE_INGREDIENT'] = {
    'abbrev': 'ai',
    'id': 'P3781'}

edge_info['HAS_CAUSE'] = {
    'abbrev': 'hc',
    'id': 'P828'}

edge_info['HAS_PART'] = {
    'abbrev': 'hp',
    'id': 'P527'}

edge_info['MOLECULAR_FUNCTION'] = {
    'abbrev': 'mf',
    'id': 'P680'}

edge_info['NEGATIVE_DIAGNOSTIC_PREDICTOR'] = {
    'abbrev': 'nd',
    'id': 'P3357'}

edge_info['NEGATIVE_THERAPEUTIC_PREDICTOR'] = {
    'abbrev': 'nt',
    'id': 'P3355'}

edge_info['PHYSICALLY_INTERACTS_WITH'] = {
    'abbrev': 'iw',
    'id': 'P129'}

edge_info['POSITIVE_DIAGNOSTIC_PREDICTOR'] = {
    'abbrev': 'pd',
    'id': 'P3356'}

edge_info['POSITIVE_THERAPEUTIC_PREDICTOR'] = {
    'abbrev': 'pt',
    'id': 'P3354'}

edge_info['SUBCLASS_OF'] = {
    'abbrev': 's',
    'id': 'P279'}

edge_info['SYMPTOMS'] = {
    'abbrev': 'sy',
    'id': 'P780'}

edge_info['MEDICAL_SPECIALTY'] = {
    'abbrev': 'ms',
    'id': 'P1995'}

edge_info['SIGNIFICANT_DRUG_INTERACTION'] = {
    'abbrev': 'di',
    'id': 'P769'}

edge_info['AFFLICTS'] = {
    'abbrev': 'af',
    'id': 'P689'}

edge_info['HAS_ROLE'] = {
    'abbrev': 'r',
    'id': 'P2868'}

edge_info['THERAPEUTIC_AREA'] = {
    'abbrev': 'ta',
    'id': 'P4044'}

edge_info['ANATOMICAL_LOCATION'] = {
    'abbrev': 'l',
    'id': 'P927'}


node_abbreviations = {k: v['abbrev'] for k, v in node_info.items()}
edge_abbreviations = {k: v['abbrev'] for k, v in edge_info.items()}

node_abv_to_full = {v: k for k, v in node_abbreviations.items()}
edge_abv_to_full = {v: k for k, v in edge_abbreviations.items()}

## Generate queries for the edges

import sys
sys.path.append('../../hetnet-ml/src/')
import graph_tools as gt

def generate_instance_tag(qname, n_id):
    return "\n    {n} wdt:P31|wdt:P279 wd:{n_id} .".format(n=qname, n_id=n_id)

def to_query_name(name):
    return '?'+name.lower().replace(' ', '_').replace('-', '')

def build_query_from_abbrev(abbrev, target=None):
    # Split the edge abbreviation in to subject, predicate, object
    s_abv, p_abv, o_abv = gt.parse_edge_abbrev(abbrev)
    
    # Get the full versions of names
    s_name = node_abv_to_full[s_abv]
    p_name = edge_abv_to_full[p_abv]
    o_name = node_abv_to_full[o_abv]
    
    # Sanitize names for query purposes
    s_qname = to_query_name(s_name)
    o_qname = to_query_name(o_name)
    
    # Need to differentiate start and end if self referential edge
    if s_qname == o_qname:
        s_qname += '1'
        o_qname += '2'
    
    # Get the proper return types
    query_text = "SELECT DISTINCT {s} {s}Label {o} {o}Label".format(s=s_qname, o=o_qname)
    query_text += "\nWHERE {"
    
    return_types = {o_name}
    
    # Build the appropriate info to get the correct subject and object node-types
    # Use set so not repeated in cases of self-referential edges
    for name, qname in [(s_name, s_qname), (o_name, o_qname)]:
        if node_info[name]['id']:
            query_text += '\n\n    # Initial typing for {}'.format(name)
            query_text += generate_instance_tag(qname, node_info[name]['id'])
            return_types.update({name})
    
        # Get any qualifiers for subjects and objects
        for qual in node_info[name].get('qualifiers', []):
            query_text += '\n    # Qualifier for {}'.format(name)
            query_text += generate_qualifier(qual, return_types=return_types)
            # Ensure there isn't the same nodetype definition multiple times
            # E.G. ?gene wdt:P31|wdt:P279 wd:wd:Q7187 # ?gene subclass/instace of Gene
            return_types.update({qual['s'], qual['o']})
                
    # Allows for changing of the target of an edge... for example if we want 
    # Sequnce Variant - variant of - Protein... SV's only go to Genes, so we can use 
    # SV - v - Gene - enc - Pro to get the relationship... 
    # Qualifiers in the protein definito handle the Gene-enc-Pro relationship, but the Target
    # variable handles the fact that SVs point to genes
    if target:
        o_qname = to_query_name(target)
    
    # The actual edge we're interested in
    query_text += "\n\n    # Edge of interest {s} {p} {o}".format(s=s_name, p=p_name, o=o_name)
    
    # Some edges are bi-directonal e.g. drug-used-for-treatment and medical-condition-treated
    # Get edges in eitehr dircetion
    rev_id = edge_info[p_name].get('rev_id', None)
    if rev_id is not None:
        query_text += "\n    {{ {s} wdt:{p} {o} }}".format(s=s_qname, p=edge_info[p_name]['id'], o=o_qname)
        query_text += "\n    UNION {{ {s} wdt:{p} {o} }}".format(s=o_qname, p=rev_id, o=s_qname)
    else:
        query_text += "\n    {s} wdt:{p} {o} ".format(s=s_qname, p=edge_info[p_name]['id'], o=o_qname)

    
    # Make sure labels return in english
    query_text += '\n\n    SERVICE wikibase:label { bd:serviceParam wikibase:language "[AUTO_LANGAGE],en" }\n}'

    return query_text
    
    

def generate_qualifier(qual, return_types=None):
    node_q_names = {}
    
    for node in ['s', 'o']:
        node_name = qual[node]
        if node_name in node_info.keys():
            node_q_names[node] = to_query_name(node_name)
        elif node_name is None:
            node_q_names[node] = '?na'
        else:    
            node_q_names[node] = 'wd:'+node_name

    qual_out = ''
    
    # if we have true node-types, we need to ensure they're used
    for semtype, qname in node_q_names.items():
        if not qname.startswith('wd:') and qual[semtype] not in return_types and qname is not '?na':
            qual_out += generate_instance_tag(qname, node_info[qual[semtype]]['id'])        
    
    # Negative qualifiers need to filter out edges
    if qual['not']:
        qual_out += '\n    FILTER NOT EXISTS {{ {s} wdt:{p} {o} }} .'.format(s=node_q_names['s'], 
                                                                            p=qual['p'], o=node_q_names['o'])
    # Otherwise add the edges
    else:
        qual_out += '\n    {s} wdt:{p} {o} .'.format(s=node_q_names['s'], 
                                                    p=qual['p'], o=node_q_names['o'])
    
    return qual_out

hetnet_edges = [
    {'abbrev': 'CdiC'},
    {'abbrev': 'CtD'},
    {'abbrev': 'PPaiC'},
    {'abbrev': 'CHhcC'},
    {'abbrev': 'PWhpC'},
    {'abbrev': 'PiwC'},
    {'abbrev': 'VntC'},
    {'abbrev': 'VptC'},
    {'abbrev': 'DaP', 'target': 'Gene'},
    {'abbrev': 'DaG'},
    {'abbrev': 'DsyS'},
    {'abbrev': 'DmsMS'},
    {'abbrev': 'CHsyS'},
    {'abbrev': 'CHsyD'},
    {'abbrev': 'VndD'},
    {'abbrev': 'VpdD'},
    {'abbrev': 'VvP', 'target': 'Gene'},
    {'abbrev': 'VvG'},
    {'abbrev': 'PWhpP', 'target': 'Gene'},
    {'abbrev': 'PWhpG'},
    {'abbrev': 'PccCC'},
    {'abbrev': 'PbpBP'},
    {'abbrev': 'PmfMF'},
    {'abbrev': 'PhpPD'},
    {'abbrev': 'PhpSS'},
    {'abbrev': 'PsPF'},
    {'abbrev': 'PhpBS'},
    {'abbrev': 'PhpAS'},
    {'abbrev': 'PhpSM'},
    {'abbrev': 'PPtaD'},
    {'abbrev': 'CrCR'},
    {'abbrev': 'DlA'},
    {'abbrev': 'CHafA'},
    {'abbrev': 'CtCH'}
]

queries = [build_query_from_abbrev(**g_edge) for g_edge in hetnet_edges]

## Run the queries

import functools
from wikidataintegrator.wdi_core import WDItemEngine
from tqdm import tqdm

endpoint='http://avalanche.scripps.edu:9999/bigdata/sparql'

def parse_result_uris(result):
    for c in result:
        if 'Label' not in c:
            idx = result[c].str.startswith('http://www.wikidata.org/entity')
            result[c][idx] = result[c][idx].str.split('/', expand=True).iloc[:,-1]
    return result.drop_duplicates()

query_func = functools.partial(WDItemEngine.execute_sparql_query, endpoint=endpoint, as_dataframe=True)

def execute_sparql_query(query_text):
    # Enforce the proper column order
    col_order = query_text.split('\n')[0].split(' ?')[1:]
    return parse_result_uris(query_func(query_text))[col_order]

res = {e['abbrev']: execute_sparql_query(q) for e, q in tqdm(zip(hetnet_edges, queries), total=len(hetnet_edges))}

# How many total edges retrieved?
'{:,} raw edges'.format(sum(len(v) for v in res.values()))

## Two potnetial issues to look into...

1. Self-referential edges...  
    These edges may contain `node1 - node2` edges as well as `node2 - node1` edges
    
2. Multi or miss-typed nodes...  
    Some node IDs may show up as multiple types... especially chemical roles and chemical compounds.

node_ids = dict()
id_to_name = dict()
self_ref = set()

for e, r in res.items():
    n_types = [c for c in r.columns if not c.endswith('Label')]
    
    for nt in n_types:
        if nt.endswith('1') or nt.endswith('2'):
            node_type = nt[:-1]
            self_ref.add(e)
        else:
            node_type = nt
    
        if node_type in node_ids:
            node_ids[node_type].update(set(r[nt]))
        else:
            node_ids[node_type] = set(r[nt])
    
        id_to_name.update(r.set_index(nt)[nt+'Label'].to_dict())
    

### Fix the Self-Ref edges where forward and backward edges may be duplicated

len(res.keys())

fixed = dict()

for kind in tqdm(self_ref):
    
    # no need to worry about forward vs reverse in directed edges
    if '>' in kind or '<' in kind:
        continue
    
    # Only look at 1 kind of edge at a time
    this_edges = res[kind]
    col_names = this_edges.columns
    
    edge_ids = set()
    
    for row in this_edges.itertuples():
        # Grab the edge ID, sorting, so lowest ID first:
        #     If both 'Q00001 -- Q00002' and 'Q00002 -- Q00001' exist, effectively standarizes to 
        #     'Q00001 -- Q00002'
        edge_id = tuple(sorted([row[1], row[3]]))
        edge_ids.add(edge_id)

    start_ids = []
    start_names = []
    end_ids = []
    end_names = []

        
    for edge_id in edge_ids:
        start_ids.append(edge_id[0])
        start_names.append(id_to_name[edge_id[0]])
        end_ids.append(edge_id[1])
        end_names.append(id_to_name[edge_id[1]])
    
    fixed[kind] = pd.DataFrame({col_names[0]: start_ids, col_names[1]: start_names, col_names[2]: end_ids, col_names[3]: end_names})

# Check an arbitrary edge that was fixed
k = next(iter(fixed.keys()))
print('Edge direction standardization results for {}'.format(k))
print('Before: {:,}\tAfter: {:,}'.format(len(res[k]), len(fixed[k])))

# numver of unique node ids
'{:,} nodes (possibly some duplicats across type)'.format(sum(len(v) for v in node_ids.values()))

for k, v in node_ids.items():
    print('{} : {}'.format(k, len(v)))

total = set()
for v in node_ids.values():
    total = total.union(v) 
'{:,} unique nodes ids'.format(len(total))

'{:,} total nodes with multiple types'.format(sum(len(v) for v in node_ids.values()) - len(total))

It will be easier to address these node issues once we've got the networks formatted into the hetnet standard

## Format Nodes into Hetnet

nodes = []
for k, v in node_ids.items():
    curr_nodes = pd.DataFrame({'id': list(v), 'label': len(v)*[k]})
    curr_nodes['name'] = curr_nodes['id'].map(id_to_name)
    nodes.append(curr_nodes)
nodes = pd.concat(nodes).reset_index(drop=True)

nodes[nodes.duplicated(keep=False, subset=['id'])].sort_values('id')

label_map = {to_query_name(k)[1:]: k for k in node_info.keys()}

nodes['label'] = nodes['label'].map(label_map)

nodes.head(10)

## Format Edges into Hetnet

The original design feature was to merege `Genes` and `Proetins` acrsos the `ENCODES` edge to create one metanode `Gene` that essentially means Gene or Gene Product.  We will start that process when building the edges

edges = []

for k, v in res.items():
    if k in fixed.keys():
        v = fixed[k]
    
    
    col_names = v.columns
    keep_cols = [c for c in col_names if not c.endswith('Label')]
    col_name_map = {keep_cols[0]: 'start_id', keep_cols[1]: 'end_id'}
    
    # Replace Proteins with Genes, to merge the protein and gene metanodes
    parsed_edge = gt.parse_edge_abbrev(k)
    if 'P' in parsed_edge:
        idx = parsed_edge.index('P')
        parsed_edge = list(parsed_edge)
        parsed_edge[idx] = 'G'
        k = ''.join(parsed_edge)
    
    v = v[keep_cols].rename(columns=col_name_map)
    v['type'] = edge_abv_to_full[parsed_edge[1]] + '_' + k
    
    edges.append(v)
edges = pd.concat(edges).reset_index(drop=True)

edges['type'].value_counts()

# merge the genes and proteins in the nodes file
idx = nodes.query('label == "Protein"').index
nodes.loc[idx, 'label'] = 'Gene'

print(build_query_from_abbrev('VndD'))

nodes['label'].value_counts()

## Now Back to problem 2, nodes that are duplicated across two different Node Types

duplicated_nodes = nodes[nodes.duplicated(keep=False, subset=['id'])]['id'].unique()

duplicated_nodes

combo = gt.combine_nodes_and_edges(nodes, edges)

# See a representitive edge of each type to ensure everything looks ok...
combo.drop_duplicates(subset='type')

#take a look at some of these problematic nodes
pd.concat([combo.query('start_id in @duplicated_nodes').drop_duplicates(subset=['start_id','type']),
combo.query('end_id in @duplicated_nodes').drop_duplicates(subset=['end_id','type'])])

# examining the above results, it looks like this is the mapping we desire
resolved_multiples = {
    'Q414964': 'Gene',
    'Q127060': 'Compound', 
    'Q7860879': 'Biological Process',
    'Q408089': 'Chemical Hazard'
}

id_to_label = nodes.set_index('id')['label'].to_dict()

nodes['label'] = nodes['id'].map(lambda i: resolved_multiples.get(i, id_to_label[i]))

print(len(nodes))
nodes = nodes.drop_duplicates().reset_index(drop=True)
print(len(nodes))

combo = gt.combine_nodes_and_edges(nodes, edges)

combo.head(2)

combo['edge_abv'] = combo['type'].apply(lambda t: t.split('_')[-1])
combo['actual_start'] = combo['edge_abv'].apply(lambda a: node_abv_to_full[gt.parse_edge_abbrev(a)[0]])
combo['actual_end'] = combo['edge_abv'].apply(lambda a: node_abv_to_full[gt.parse_edge_abbrev(a)[2]])

bad_edge = combo.query('start_label != actual_start or end_label != actual_end')
bad_edge

# Becasue combo is made by merging the nodes onto the edges, the index is constant between the two DataFrames
edges = edges.drop(bad_edge.index).reset_index(drop=True)

## Write out results

out_dir = '../2_pipeline/00_hetnet_design_and_build/out'

if not os.path.exists(out_dir):
    os.makedirs(out_dir)

gt.add_colons(nodes).to_csv(os.path.join(out_dir, 'nodes.csv'), index=False)
gt.add_colons(edges).to_csv(os.path.join(out_dir, 'edges.csv'), index=False)

json.dump(node_info, open(os.path.join(out_dir, 'node_info.json'), 'w'), sort_keys=True, indent=2)
json.dump(edge_info, open(os.path.join(out_dir, 'edge_info.json'), 'w'), sort_keys=True, indent=2)
json.dump(hetnet_edges, open(os.path.join(out_dir, 'hetnet_edges.json'), 'w'), sort_keys=True, indent=2)
