==> 1_code/01_get_gold_standard.ipynb <==
import os
import requests
import pandas as pd

rel = pd.read_csv('../0_data/manual/drugcentral_rel_06212018.csv')
syn = pd.read_csv('../0_data/manual/drugcentral_syn_06212018.csv')
ids = pd.read_csv('../0_data/manual/drugcentral_ids_06212018.csv')

rel.head()

syn.head()

ids.head()

ids['id_type'].unique()

ids['id_type'].value_counts()

import functools
from wikidataintegrator.wdi_core import WDItemEngine
from tqdm import tqdm

endpoint='http://avalanche.scripps.edu:9999/bigdata/sparql'

def parse_result_uris(result):
    for c in result:
        if 'Label' not in c:
            idx = result[c].str.startswith('http://www.wikidata.org/entity')
            if sum(idx) != 0:
                result[c][idx] = result[c][idx].str.split('/', expand=True).iloc[:,-1]
    return result.drop_duplicates()

query_func = functools.partial(WDItemEngine.execute_sparql_query, endpoint=endpoint, as_dataframe=True)

def execute_sparql_query(query_text):
    # Enforce the proper column order
    col_order = query_text.split('\n')[1].split(' ?')[1:]
    return parse_result_uris(query_func(query_text))[col_order]

query = """
SELECT DISTINCT ?umlscui ?disease
WHERE {
  ?disease wdt:P2892 ?umlscui .
  SERVICE wikibase:label { bd:serviceParam wikibase:language "[AUTO_LANGAGE],en" }
}
"""

res = execute_sparql_query(query)
res.head(2)

len(res)

umls_to_wd = res.set_index('umlscui')['disease'].to_dict()

rel['disease_wd'] = rel['umls_cui'].map(umls_to_wd)

rel.query('relationship_name == "indication"')['umls_cui'].count(), rel.query('relationship_name == "indication"')['disease_wd'].count()

struct_to_chembl = ids.query('id_type == "ChEMBL_ID"').set_index('struct_id')['identifier'].to_dict()

rel['comp_chembl'] = rel['struct_id'].map(struct_to_chembl)

rel.query('relationship_name == "indication"')['comp_chembl'].count()

query = """
SELECT DISTINCT ?chembl ?compound
WHERE {
  ?compound wdt:P592 ?chembl .
  SERVICE wikibase:label { bd:serviceParam wikibase:language "[AUTO_LANGAGE],en" }
}
"""

res1 = execute_sparql_query(query)
res1.head(2)

len(res1)

chembl_to_wd = res1.set_index('chembl')['compound'].to_dict()

rel['comp_wd'] = rel['comp_chembl'].map(chembl_to_wd)

rel.query('relationship_name == "indication"')['comp_wd'].count()

gs = rel.query('relationship_name == "indication"').dropna(subset=['comp_wd', 'disease_wd'])
len(gs)

out_dir = '../2_pipeline/01_get_gold_standard/out'
if not os.path.exists(out_dir):
    os.makedirs(out_dir)
    
gs.to_csv(os.path.join(out_dir, 'gold_standard.csv'), index=False)
